Fix error allowing multiple subscribers with same id to same path
Summary:
OperSubscriberInfo is used as a unique key for our active subscriber set. This set is then used to determine wether a subscriber is a duplicate or not.

With the addition of subscribedSince in D50436500, we've been allowing subscribers with the same id and same path, because the time is always unique. Introducing an new ClientKey to get around this.

In the future I'd like to get rid of this map in favor of the info inside SubscriptionManager as a SoT, but this will do for now.

Differential Revision: D55086021

fbshipit-source-id: ed265dfacd012fb427dc411a469ccebce8d73efc