Adding CRC 16 calculation for weutil print
Summary:
Will print if CRC mismatch is found when using weutil to dump eeprom.

Note that the EEPROMs are not programmed correctly on BSP platforms. From now on, we recommend using Crc16CcittAug as the path forward for calculating the CRC.

For Darwin, all the units that I looked at in prod / lab have 0x0 programmed in place of CRC.

Test Shows the mismatch.

Reviewed By: somasun

Differential Revision: D54301669

fbshipit-source-id: 7795bec166773eeee410dfc41b20aa32a5b5ac4c