Option to control initial sync response behavior
Summary:
D54819333 introduced behavior of always sending a response (even if empty) on
initial sync to subscription stream from CowStorage server.
CowStorage is common to FSDB and NSDB, but this behavior is only required of
FSDB protocol. NSDB will investigate and if needed enable this behavior.

This diff adds an option to enable this initial sync response behavior,
and enables this behavior only for FsdbCowSubscriptionManager.

Reviewed By: peygar

Differential Revision: D55106552

fbshipit-source-id: 8ab3b48f925422326b5d8906bb395e1b3e5732d4