Don't crash cow storage when delta path is invalid
Summary:
CPS services use subscribable storage but don't have a filter for invalid deltas like FSDB ServiceHandler does. As such, we're seeing occasional crashes (P1198522805) when NSDB is ahead of CPS and sends unrecognizable deltas.

Adding some extra safety at the library level to not crash on invalid paths. During the modify step we'll know if the path doesn't exists. In those cases we end up not modifying the tree, so trying to traverse writable and patch will throw an error. Changing to exit early if the modify failed and return a not found error

Reviewed By: wilsonwinhi, Linerd

Differential Revision: D55147546

fbshipit-source-id: 839d5d3c770057744cf5f8cb06b3b2a27cb70330