Add an option to skip stats update from reading from HW for debugging
Summary:
For drop related debugging with ASIC vendors, it would help to have all the stats be
available via diag shell. FBOSS stats thread running at this time would mean a lot of
read-on-clear counters will be cleared and not available when checked from diag shell.
To avoid this, introducing an option to avoid the updateStats thread from reading from
HW as a config knob which can be used from thrift.

Reviewed By: msomasundaran

Differential Revision: D55204712

fbshipit-source-id: 01f26e33048d55769edbfd85a6320ede881a295a