Add command to show fabric topology per virtual device
Summary:
Command used to figure out asymmetric topology
connections. In prod we expect only a single FAP device
to connect to a FE virtual device. In labs however we
violate this constraint. But even when multiple links
connect, it should always be symmetric. That is the
same number of links should connect from each FE
virtual device to all FAPs connecting to that virtual
device.

Reviewed By: nivinl

Differential Revision: D55298939

fbshipit-source-id: 2275ea3539b28e2c11bdfe4dc7a6f7da8e00ddef