Use Type=notify in platform_manager.service
Summary:
PM will remain exploring only once (for now). But the binary will be active and run thrift server in ssd. But remain the same in ramdisk. X86 services will only run after PM completes; this could mean both on success and failure

For example, if PM fails (timeout or abort) on the first run and proceeds to retry, other services will start without waiting. This behaviour is consistent with the current oneshot behavior.  I think we should enforce more stricter ordering (or even dependencies). But I think this can be addressed separately.

Reviewed By: somasun

Differential Revision: D54553370

fbshipit-source-id: 9ab014904a8640db841527f571fd851f4a6164a5