Simplify lock contention
Summary:
The current sensor_service locks the entire data-struct while
performing sysfs reads. When the number of sensors are large, this takes longer
than 5s (default sensor fetch interval). This results in the lock being held
for a long duration. If a thrift call for sensor-data happens during this time
(which also is acquiring lock), it can result in timeout.

This diff simplifies the critical section to involve just copy of polled data.
The lock is not held on the data-struct for the entire duration of polling. We
have to look at parallelizing the sysfs reads (for large number of sensors), so
we complete the sensor reads within 5s.

Reviewed By: kimdo8736

Differential Revision: D55245680

fbshipit-source-id: 962f2add9ca598c35517e3c76e4613d73c6d82db