Read power control from hardware before clearing low power
Summary:
When configuring 8x100G mode, it is possible that we do back to back 8 customizeTransceiverLocked call. In each call, we try to see if the cached power control is high power mode. If it is not in high power mode, we first enable low power, then clear it.

When 8x calls happen back to back, there is no time in the middle to update the cache. Thus we set and reset lpMode 8 times instead of once.

In this diff, i modified the code so that we always read the powerControl from hardware before deciding if we should toggle low power or not. This seems safer than relying on the cache to be always updated.

Reviewed By: rajank7

Differential Revision:
D55408547

Privacy Context Container: L1125642

fbshipit-source-id: 551a9c9874c1a8f479e725f40b3f5475a2e5307c