Default isAttached to false.
Summary:
Default FabricEndpoint carries garbage connectivity info
switchid=0, portId = 0, switchName=portName= "". So set the default
to not attached as this is not the port or switch we are attached to.
Note that FabricEndpoint gets inserted into the map as soon as
port is added so its important to have the right values here.

Reviewed By: shri-khare, zechengh09

Differential Revision: D55541412

fbshipit-source-id: 03acac5a692a42116d0b486635b041337e740b73