Disable variable name logging in prod
Summary:
This is towards supporting Scott's parser on replayer logs to analyze and identify routing loops.

As a first step, the variable naming in prod device are making it harder to parse (as it requires to go back to older logs to find the corresponding object.

Instead of logging with variable names, log directly object IDs - this enables replayer logs easily work with the object IDs return with `listHwObject`.

This also reduces memory usage in prod.

D55565992 to enable this flag by default for sai test/sai bench and sai link test config.

Reviewed By: nivinl

Differential Revision:
D55464515

Privacy Context Container: L1125642

fbshipit-source-id: 71667d546546507798e4d6a80a0af0820e89fd53