Traverse children first when serving subs
Summary:
For Patches the plan is to serve patch subs as we're traverse, same as other subs. To do that, we need the Patch sub tree to be ready the moment we visit a node. So using a little trick here to traverse CHILDREN_FIRST, allowing us to prep the Patch before getting to a subscription. Then we can just copy over our Patch into the pipe.

This shouldn't cause any issues for current subscriptions, but split this into a dedicated diff to get UT signal, and in case we need to revert

Differential Revision:
D49109972

Privacy Context Container: L1125642

fbshipit-source-id: 320d0fcbdc88bedfa184cb540812a0ea1337073f