Always send link connectivity delta event in linkConnectivityThread
Summary:
This is needed to synchronize with initial link connectivity
communication. When the swagent/hwagent (re)connect we communicate
the entire (current) connectivity upto the sw agent. Following which
deltas are communicated. By doing this in a common event base and
given the fact that on connection
- we first do the processing for connected (where we will do the
initial sync)
- followed by event processing (where subsequent connectivity deltas
will be queued by stats thread)
https://www.internalfb.com/code/fbsource/[04b5ab8643a69ecbffaf194198eb5393c174ac01]/fbcode/fboss/agent/mnpu/SplitAgentThriftSyncerClient.cpp?lines=145%2C149
We are guaranteed that full sync will happen before any event
communication.

Reviewed By: srikrishnagopu

Differential Revision: D55724745

fbshipit-source-id: 410e49c1b9afaad868777141ecca34957d4bd86c