configure unresolved route to acl action to trap
Summary:
For Cisco devices, two packets are punted to CPU as part of UnresolvedRouteNextHopToLowPriQueue test.

This is because packet hits the ACL is copied to CPU and then also hits the MYIP trap and is trapped to CPU again. With the ACL in place, the packet must be trapped to CPU via queue 0 and not reach mid pri queue. S390808 for more details.

For unresolved route to CPU ACL, explicitly configure the trap action as TRAP instead of COPY.

Differential Revision: D55819540

fbshipit-source-id: a1ee408c62812d3ecd70a8f3c4c6b8fc8ebaaf4e