Don't queue a disable port update if port is already disabled
Summary:
As titled. Noticed that in
AgentFabricSwitchSelfLoopTest.selfLoopDetection we were querying
multiple updates for the same port. This is because we first
queue updates to disable all the ports. Each of those then
trigger loop detection and trigger updates for all other
ports. This leads to N^2 updates being queued even though
only the first N are needed

Reviewed By: srikrishnagopu

Differential Revision: D55833326

fbshipit-source-id: 71086d61182a3d3395b64eeae44bdf2399b84f2d