Split SubscriptionManager non-templated methods to base class to move to cpp
Summary:
The goal of this stack is split as many methods that don't depend on templates into their own translation units. This won't help too too much in terms of clean compile times, but it should help quite a bit with incremental builds while making changes to these libraries, if the change is limited to these non-templated methods. Instead of having to build the entire library including all the visitors and cow storage, we'd only rebuild the one non-templated cpp file.

This diff handles splitting subscription manager. Next diffs tackle subscribable storage.

I'm doing this now because I'll have to iterate on these libraries a bunch while looking for potential memory wins. Hoping this investment will pay dividents by the end of this excercise, but if not will eventually be worth it

Differential Revision: D55704321

fbshipit-source-id: 2c138f0f7fa2d34f57f632a146166827b611ef40