Cleanup flag state leftover by getAclTableGroup thrift test
Summary:
This test sets a command line flag, but doesnt reset it. That means
if the next test is run in the same process the flag will remain set
and may cause failures.
We were getting lucky with default buck/sandcastle runs since those
always run a individual test in a separate process

Differential Revision: D55858768

fbshipit-source-id: 90164917ed5f89018f6b31f71ca59f885e698dc4