Move sub mgr related methods to base
Summary:
These methods only use methods in SubscriptionManagerBase (non-templated version). So it's possible to move them all to the non-templated NaivePeriodicSubscribableStorage too.

Only trouble is because we have this wrapped in a folly::Synchronized which is not trivially convertable to the base https://fb.workplace.com/groups/474291069286180/posts/26030163186605617.

To get around this I had to do a little bit of gymnastics with some wrapper classes

Differential Revision: D55704322

fbshipit-source-id: 811f4e22e1eb5ed0ddb68ed9abaeaf8fd38986ab