Thrift changes for DLB selective feature
Summary:
As described in the doc, DLB requires matching on the 16th byte offset in the ROCE header. In addition, the match also requires a mask.
This change introduces 2 new fields, roceBytes and roceMask. They will eventually deprecate the existing roceOpcode field.
The new fields are also lists to allow more flexibility in how many bytes are matched in the future.

Differential Revision: D55938427

fbshipit-source-id: 8180571b2156b36507371397bd01fa6894a24d3d