Simplify fan presence/rpm/pwm/led access
Summary:
All of the platforms use sysfs currently. rpm/pwm/led will always be sysfs
based. Presence can be gpio based in the future (which is not
supported/implemented in fan_service yet). When we have a use-case for that, we
can extend the Fan thrift struct to include that information. We can
incorporate validation that fan presence has either the sysfs or the gpio
handle specified (not both)
Also, simplified the logging.

Reviewed By: mikechoifb

Differential Revision:
D56006130

Privacy Context Container: L1125642

fbshipit-source-id: 78a6a68caceb773d0b482e4b33676eb9404683eb