ReadEeprom Library
Summary: Adds a library to read directly from an i2c device and output the results into a file. This will be used by weutil and PlatformManager to handle a bug which makes it impossiblle to read the Meru SCM eeprom regularly. See stacked diffs for testing and more context.

Reviewed By: somasun

Differential Revision: D55612853

fbshipit-source-id: 82e3d04ea5c234718ee3e1832e3a504cc7643b6f