Handle P2 DSF eeprom issue
Summary:
Because of an issue with an upstream driver, DSF P2 units need special handling of the SCM eeprom. Details here: https://github.com/facebookexternal/fboss.bsp.arista/pull/31
(TLDR if you don't have access: upstream kernel driver assumes that i2c-1-005* is a DIMM and binds the ee1004 driver to it, and we are unable to unbind it afterwards).

PlatformManager is adjusted to read directly from the i2c device when detecting the meru scm eeprom. It writes direcltly to the previously symlinked path.

Reviewed By: rajank7

Differential Revision: D55658224

fbshipit-source-id: 1727516b42a0a7f5f104bff509d2789a55fd59a5