Make module level get symbol error function thread safe
Summary:
Made the function to obtain the Symbol Error histogram thread safe so that the actual function doing HW reads is executed in corresponding event-base only. This is because this function execution may be triggered from thrift calls also.

NOTE: I am not planning to call this function from refresh cycle. This function to read Symbol error histogram is stateful which means we issue the CDB command and wait for some time for CDB to respond. Such call will take more time and we can't put this in refresh cycle. So this function is currently getting called from thrift thread only

Reviewed By: birdsoup

Differential Revision: D55767480

fbshipit-source-id: 64760ec6144b582cb09fbcc03fb2cc24ff6d409e