Fix fabric utils check reachability api
Summary:
This check and test was failing since the time
we added virtual devices concept. The problem is that
we are trying to lookup the switch based on switch id
recieved in reachability info from HW. This works fine
on devices without virtual devices. But on R3 the swithid
in reachability would factor in the virtual device id.
So for R3 you would get
switch Id and switch Id + 1
However only switchId exists in the ASIC table, dsf node
map etc.

For expected switch type - there is a simple solution in
tests. We already know our switch type, and ports are in
loopback mode - so just use that.

Reviewed By: nivinl

Differential Revision: D56371929

fbshipit-source-id: 6cae1578d2a27f82dcce050f35c3cbfe752e71ce