Olympic qos helper - accept multiple asics but assert that they are of same type
Summary:
As titled. Towards making the tests not assume a singe ASIC.
Individual utils may assert that asics must be uniform, but tests
need to be written to think in terms of multiple ASICs.
For Qos utils - for now assuming uniform asic types. Later this
could be relaxed to be compatible ASIC types -e.g. J3 and J4.

Reviewed By: phshaikh

Differential Revision: D56406872

fbshipit-source-id: 00e825da28ae028956a81c3b0f4fafc2f4a390b6