Implement incremental compression
Summary:
Benchmarking from last diff revealed something I had feared for a while. The overhead from containers (maps) needed to store a patch tree turns out to outweigh the memory savings we'd get from condensing the paths. Luckily theres a relatively simple fix to this that makes this consistently beat deltas in both memory and serialized size by a significant factor.

Here I'm introducing a concept of incremental compression of the patch subtrees. Any time we're walking back up the tree, it means we're done adding modifying that subtree. So we can "compress" that subtree by preemtively serializing it and storing it as a binary instead of a bloaty in memory map. During patch application, we just have to make sure to deserialize the compressed children before trying to apply

Differential Revision:
D56270813

Privacy Context Container: L1125642

fbshipit-source-id: 504b62d09f2439bc2bd1d93f956acb4ce4042901