fboss/platform_manager: allow to setup i2c registers in thrift
Summary: This diff introduces a new thrift structure called `I2cRegData` which allows the setup of I2C registers in thrift. The `I2cRegData` structure contains the register offset and the data to be written to the register. This new structure is added to the `platform_manager_config.thrift` file. Additionally, the `I2cDeviceConfig` structure is updated to include the `I2cRegData` structure. This update allows the platform manager to set I2C registers in thrift.

Reviewed By: somasun

Differential Revision: D55549477

fbshipit-source-id: 06a04854bb9d1b596bd2095888b6778af9cd8946