fboss/platform/configs/janga: setup adc128 operation mode in pm.json
Summary: This diff sets up the operation mode for the ADC128 in the platform manager configuration file for the Janga800BIC platform. The code changes include adding an "initRegSettings" section to the "adc128d818" device in the platform manager configuration file. This section sets the initial register settings for the ADC128 to enable the operation mode.

Reviewed By: somasun

Differential Revision: D55549478

fbshipit-source-id: 90490f965b911c047045561423403f0d47c8cd51