New apis support patch and other features
Summary:
Our chunk type will be changing with these new patch types, so we're going to need new apis. As part of this, I'm taking the liberty to make a few extra changes based on our learnings

1. Stream chunks are now "Message" unions to allow for heartbeats and other potential future control chunks
2. Making the default subscription apis multi-path friendly:
a. Subscription request now takes a `{key -> path}` mapping. The key for each path is an integer defined by the client
b. The subscription chunk contains the key which the client can use to efficiently figure out which path is being served, rather than having to do string comparison against the path. We could have done list<Path> and have the key just be the index into the list, but to me this seemed less error prone around adding/removing paths from a subscription
c. For C++ all this logic will be hidden behind client libraries
3. To support a "Path"-like subscriptions, adding a flag serveShallowPatches. With this client will receive non-granular Patches at the subscribed path. Adding this so we can get rid of both delta and path subscriptions.
4. Getting rid of extra keywords like Oper, Path, Delta from apis since people outside the team did not really know what they meant. We can probably get away with just subscriberStat(e|s)
5. Changing protocol default to compact, generally recommended nowadays

Differential Revision: D49746101

fbshipit-source-id: e9c5f8c394c788b2dba17699cace5f11ad11c54c