Remove unused variables in fboss/agent/hw/bcm/BcmIntf.cpp
Summary:
LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but, eg, in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: jasmeetbagga

Differential Revision: D56587591

fbshipit-source-id: 09eadbaa24aa16a3b06c6eb5391c51f02b4ae958