Fix Wmismatched-tags in fboss/agent/platforms/sai/SaiPlatformInit.h + 5
Summary:
`struct` and `class` can be used interchangeably in some instances in C++, but not all!

The two mean different things with regards to member privacy settings and, in MSVC, interchanging them can lead to ABI issues.

This diff switches `struct` and `class` as appropriate to make the codebase self-consistent. Doing so satisfies the `-Wmismatched-tags` flag.

#buildsonlynotests - If this diff compiles, it works.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D56615869

fbshipit-source-id: 9bc5612c68c0b155d03e6ff4ee627fa61b043d2d