Get supported port profile id function at TransceiverManager level
Summary: Added a function to get all supported port profiles at the transceiver manager level. If the passed argument "checkOptics" is false then it will return all possible port profiles for that port. If the "checkOptics" is true then it will exclude the port profiles which are not supported by the current optics

Reviewed By: harshitgulati18

Differential Revision:
D56507875

Privacy Context Container: L1125642

fbshipit-source-id: 5f5745215c358aebeb1c137b3c9e02de61072338