Fix Wmismatched-tags in dolores/gateway/DataPlaneLogging.h + 5
Summary:
`struct` and `class` can be used interchangeably in some instances in C++, but not all!

The two mean different things with regards to member privacy settings and, in MSVC, interchanging them can lead to ABI issues.

This diff switches `struct` and `class` as appropriate to make the codebase self-consistent. Doing so satisfies the `-Wmismatched-tags` flag.

#buildsonlynotests - If this diff compiles, it works.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D56615864

fbshipit-source-id: 5893ef8cbd5cd2588d9863d90da79d4ffd9cd21e