Implement Handlers for Bulk Clear Interface PRBS Stats HwAgent Thrift API
Summary: I implemented the bulk clear interface PRBS stats Thrift API on both BcmHandler and SaiHandler. In either case, we iterate through the interfaces, check if HwAgent is able to recognize current interface. If it recognizes, we call the HwSwitch handler to clear PRBS stats for that interface. If not, we skip.

Reviewed By: zechengh09

Differential Revision:
D56369485

Privacy Context Container: L1125642

fbshipit-source-id: e3d2b0db6ad9015438f68c6223c07d3dca91978f