Fix led offsets in PM config
Summary: The led offsets were messed up in the PM config so 1-32 were off by one, and 33-64 were all completely missing. It appears the formula for generating the old config values was messed up. This diff fixes them and also adjusts the ordering in the config just because it was easier to make a script that output them in sequential order.

Reviewed By: rajank7, xanabani

Differential Revision: D56729898

fbshipit-source-id: 7cd665770129feffc73f1ee0cf05209736aed988