Ensure the pwmLower and pwmUpper thresholds are honored
Summary:
Previously, the pwmLower and pwmUpper thresholds were checked in
getSensorUpdate() function.  There was no guard against pwm going outside the
range in the subsequent workflow (like calculateZonePwm). This could result in
pwm going out of the range. In this diff we move this check to just before
programming pwm (in programFanPwm)

Reviewed By: tao-ren

Differential Revision: D56943354

fbshipit-source-id: c94abc2dc4d7ab2d54cb81f448a6f1464dcdc7fb