Optimize PRBS Stats Caching when Enabling/Disabling PRBS
Summary:
If PRBS is disabled on a port, we want PRBS handlers to be as close to no-op on that port as possible.

When we add port, we check if PRBS is already enabled on port. If so, we add stats entry for that port to the cache. When we change port, we check if PRBS state changed. If it did, if new state is enabled, we add stats entry. If new state is disabled, we remove stats entry. Removing port will remove stats entry from cache regardless.

The get PRBS stats and clear PRBS stats handlers will first fetch PRBS state on the current port. If the state is disabled, the handlers will skip that port and move on to the next port (since that port would not have an entry in the cache).

Reviewed By: zechengh09

Differential Revision:
D56656262

Privacy Context Container: L1125642

fbshipit-source-id: 8df6b70f8ee086a2ca63010f37b05695ecd0e623