Query HwAgent for FBOSS2 Clear Interface PRBS ASIC Stats
Summary:
I refactored the implementation for fboss2 clear interface prbs stats such that it utilizes HwAgent for ASIC PRBS. We leverage the same bulkClearInterfacePrbsStats API for all components, including those managed by QSFP.

To determine which interfaces to pass to which HwAgent, I make a Thrift call to SwAgent to get switch indices for interfaces. Based on this result, we should be able to pass the correct interfaces to the correct HwAgent Thrift call.

Reviewed By: zechengh09

Differential Revision:
D56443472

Privacy Context Container: L1125642

fbshipit-source-id: 11e024b1d01df071450b3b17adb09d56d897d11f