Create FsdbSyncer earlier, start it later
Summary:
Right now we don't construct FsdbSyncer until initialConfigApplied, so any updates before then will be silently dropped. For switch state and config we explicitly set the initial state in FsdbSyncer::FsdbSyncer, but we were not doing that for fsdbSubscriptions. The effect is a race between dsf subscriptions and initial config application, any streams that converge before agent is done with initial config will have the session state desynced.

Here I'm changing the setup to create FsdbSyncer right away, but dont `start` it until after initialConfigApplied. This way we never end up missing updates that are queued to FsdbSyncer, but we only actually start publishing when our switch state is populated. This model will help avoid this type of bug in the future

Reviewed By: shri-khare

Differential Revision:
D57139989

Privacy Context Container: L1125642

fbshipit-source-id: 6a6b89b3057ea2ac5c589a79fd92a6121dea1ae8