Collect sysPortStats
Summary:
This diff is related to the collection of sysPortStats in Netstate. The code changes in the diff include adding a new function to subscribe to sysPortStats in the fsdb_cgo.cpp file and updating the fsdb_cgo.h file to include the new function. The purpose of this collection is to monitor the status of system ports in network devices.

We need these counters to monitor the SAI_QUEUE_STAT_CREDIT_WD_DELETED_PACKETS (watchdog deleted packets), as per https://fburl.com/gdoc/hnkrgv7g

Reviewed By: Reincarnationist

Differential Revision: D55600101

fbshipit-source-id: 94dd171a14ce1e9b8f408614e65e39ddef4ab814