fan_service: Implement stub `setHold` and `getHold` thrift methods
Summary:
These methods will be called from the fboss cli to set a hold on the
fan PWM (i.e., set the PWM to a constant value that will not change with
future sensor readings). In this diff, the set and get methods only set and get
a variable and do not affect the fans. This will be fixed in future diffs.

Reviewed By: mikechoifb

Differential Revision: D57470318

fbshipit-source-id: 597995bcbcc33e367c13f2d7ab482f6c5fd6022a