ICMPv4 Interface and Next-Hop Identification
Summary:
## Intent

As we're designing the address allocations and routing for DCTypeF and future topologies, we would like to do away with the requirement of any form of IPv4 loopback address on the intermediate routers yet still give us the ability to route and troubleshoot V4 prefixes.  Today, we can use RFC5549 to carry V4 prefixes over V6 next-hops, but we keep the V4 address on the loopback for the purpose of traceroute troubleshooting.

This diff implements RFC4884 "Extended ICMP to support multi-part messages" and RFC5837 "Extending ICMP for Interface and Next-Hop Identification" behavior such that a TTL Exceeded message will contain the IP Address of the interface on which the datagram arrived.

## Additional ICMP complexity
Today, we're following the simple ICMP header and 8 bytes of the datagram payload.  This diff changes this to suite RFC5837, will put as much of the original payload as possible without going over the reassembly size

This diff enforces extension headers with the IP Address Sub-Object.  This should also be configurable and we should implement the other identification mechanisms.

This diff makes part of rfc5837 section 4.3 for our use-case to identify the interface AND router:
> The interface name MAY be some other human-meaningful name of the interface

Reviewed By: phshaikh

Differential Revision: D48922763

fbshipit-source-id: 6fd35edf857869c9169750f2d66ef45bc8033ba2