Modify queue shaper test to look for an upper bound as well
Summary:
We expect shaper to limit traffic to a specific rate. As of now, we are ensuring the rate
is >= what we expect, however, there is no upper bounds check. Though upper bounds
really vary based on port speeds and the shaper granularity, adding an upper bound of
4% for now as it works. We dont expect this to change as that can result in hosts
potentially dropping traffic due to the PCI limit.

Context: S416694 concern if TAJO is actually rate limiting as expected on MHNIC.

Reviewed By: srikrishnagopu

Differential Revision: D57762096

fbshipit-source-id: ef748b09b4da44af50b927e35d719846342fe4f8