Improve log when publishers drop
Summary:
We noticed a weirdness in how we send publisher drop notifications. The scenario is

- publisher A (ex mka) has never connected
- there is a subscriber to publisher root A
- unrelated publisher B (bgp) drops
- we find all publishers that are not connected which includes A
- we send an exception to all subcribers of A

This was notificed because netstate unconditionally subscribers to mka so on devices that don't have mka we saw some streams flapping seemingly out of the blue

Changing to at least show the publisher root that we're sending an exception for to make it easier to figure out.

Note this ends up showing the field id rather than name so it's a bit hard to figure out, I think I need to write a proper translation layer to convert path ids back to names for logging

Differential Revision: D57780322

fbshipit-source-id: 100ee4cc1f57d3600d6a4bbc0276fa9de17ebdec