hw agent handling to exit sink event wait loop
Summary:
On stress testing with repeated warmboots, hw agent can get stuck trying to shut the ipc sink streams to sw agent.  The sink event generator waits on a queue for next event and enqueues the event to the sink. Normally when sw agent exits, the sink event generator gets an thrift connection failed exception and exits the loop. In failing cases, the exception is not observed and the thread waits for the next event and eventually hits SIGABRT.

Fix: Enqueue a dummy event so that the wait loop can be exited

Another option tried was to use cancellation for the dequeue call. The hw agent is able to exit in this case but if sw agent alone restarts, then event waiting code is cancelled and do not get an exception which allows it to retry connection to sw agent

Differential Revision: D57868686

fbshipit-source-id: 418df61a6512327b89568c4a1788a58a6372da9b