Add FBPKG config
Summary:
Pytest discovers tests in a specific directory that the user configures. In our case this is through the main function with `pytest.main(..., <test_directory>)`. This means tests files must be installed in the correct place in order for this to work. Previously I was handling this with `setup_switch.py` which works for development, but is kind of hacky.

This adds an FBPKG config to the TARGETS file to enable building this fbpkg. Package name and ACL have already been created in other diffs.

fpkg link: https://www.internalfb.com/fbpkg/package/fboss_bsp_tests?showEphemerals=true

Reviewed By: somasun

Differential Revision:
D57632818

Privacy Context Container: L1125642

fbshipit-source-id: a2814215db392c21b0005c99de1f68cd5a3bf3a1