Provide helper to schedule state update in event base thread
Summary:
As titled. During the initial attempt of remote ECMP programming, we noticed the race condition of remote intf updates in RIB vs. route programming from BGP.

Imagine the following scenario
1. DsfSubscriber receives update from another RDSW, scheduling update.
a. `stateUpdate()` will take the `pendingUpdates` lock, then
b. it will try to update RIB within ribUpdateThread in blocking manner.

2. Agent receives route programming from BGP. Route update wrapper will
a. Take ribUpdateThread to compute route updates in FIB
b. `stateUpdate()`  to take the `pendingUpdates` lock.

Both threads are trying to take lock in reverse direction, hence causing deadlock.

To mitigate this, plus to support rollback of DsfSubscriber, we decided to have Dsf updates and FIB update in the same transaction. If any update failed to applied in the hardware, we would be able to rollback both of the updates.

Implementation wise, the update can be queued in ribUpdateThread, such that the lock sequence will be the same as route update (scenario 2 above).

Reviewed By: shri-khare

Differential Revision: D57918119

fbshipit-source-id: ac1d18c57c397a052a687d2b680bb778a08de3a4