Publish counters for high temp and vcc to fb303
Summary: In this diff I publish a single global counter for how many times a transceiver crosses the high warn or alarm flag and similarly for vcc. The idea behind a single counter vs per interface counter is that the use of this counter would be to just find devices which have at least one over heating transceiver. We won't have any detections based on this counter. Anything (netstate) that needs to act on per interface high temp will look at the per interface TransceiverInfo in FSDB for that.

Reviewed By: xanabani

Differential Revision:
D58048439

Privacy Context Container: L1125642

fbshipit-source-id: ea32847b84bc7365c073d6a18051027c8b832ce7