only enqueue non-noop lambda in linkConnectivityThread
Summary:
We are enqueueing a noop lambda on to an event base that
isn't actually running.

This fixes to avoid the enqueue if there is no actual delta to be
processed. For XGS switches, the delta will always be empty.

We will gate all logic around the exchanging connectivity behind the
FABRIC_PORT asic feature in a future diff, but this is a mechanical
fix that should fix the leak and have no functional change.

Reviewed By: shri-khare, srikrishnagopu

Differential Revision:
D58150754

Privacy Context Container: L1158612

fbshipit-source-id: c1aa2040e715aaab5cd80373b7e03fc4f6399ce8