Fix to contain >128 DLB ecmp groups
Summary:
With DLB, the number of supported groups on TH asics is 128. During cold boot, before routes converge, there are a large number of ECMP groups created. Any ECMP group >128 (200128 BRCM ID) is created as a regular ECMP group. It is possible that this group never gets migrated down <128 and remains "stuck" as a regular ECMP group.
This change utilizes the ResourceAccountant infra to block ECMP group creation when count goes above 128.
A caveat now is after this change, the maximum number of ECMP groups is limited to 128 or 96 if the ecmp_resource_percentage is unaltered from its default value of 75%. This doesn't pose an issue at the moment but needs revisiting if ever the requirements change. A couple of alternative approaches are proposed in https://docs.google.com/document/d/1czU7akgr4wdjbs0EQzqsiD9qiL1hYf4jn_AZTMgGTfI/edit#heading=h.nsextkmtgml.

A corner case where if the number of ecmp groups is >96 prior to warmboot (for the case where this DLB resource check is not yet rolled out), the check will stay disabled until the next warmboot. This is to ensure the DLB resource check failures wouldn't abort the agent or when agent is warmboot from a prior transient state with high number of ECMP groups. This is a precautionary check and can possibly removed in the future.

If there is ever a case where this resource check rollout leads to unintended or unpredictable outcomes, a flag is introduced to immediately disable resource checks (dlbResourceCheckEnable).

Reviewed By: zechengh09, msomasundaran

Differential Revision: D57506090

fbshipit-source-id: e40bf0edeeb0b5df68dcd52af8c5d06f5ac21a9b