Do not let clients choose patch protocol
Summary: An unfortunate side effect of incrementally compressing patches is that its no longer very efficient to allow the client to pick the encoding protocol. This is because if we have two different clients that want different protocols, we'll end up having to create two identical patch trees just with different protocols. Instead for now we'll just restrict this feature which will make it much more efficient to just copy and send parts of a single patch tree

Differential Revision:
D57055223

Privacy Context Container: L1125642

fbshipit-source-id: 1a8f917a6cbf2ee03e107d7a79bb6e61943b888c