Make sure we properly send patches on added / remove paths
Summary:
Introduced a bug here after moving to extended subscriptions. Previously when Patches were simple subscriptions, when a new path was added we'd handle it correctly in initial sync. With extended subscriptions, initial sync just resolves the subscription and we rely on the main serveSubscription loop to notice the new path and send to the subscription.

To handle this, we need to implement this similar to delta subs. Any time there is a minimal change OR added/removed path we know that the patch subscriber needs to hear about it

Differential Revision:
D57495815

Privacy Context Container: L1125642

fbshipit-source-id: 4a7d8ea9a2436c0adfe9cdd82a3d053dc5396620