Blink LED if port is down/has cable errors AND it is drained
Summary: If a LED blinks, it means it is safe to install a loopback on it and service the port. So far, the condition to not blink has been if the ports are still in active state. Certain platforms support per interface drains. So if the interfaces are not drained, it should also be a check in deciding if the led should blink or not.

Reviewed By: rajank7

Differential Revision:
D58310546

Privacy Context Container: L1125642

fbshipit-source-id: c8a86a878289802cd28ad6d62d9b71c259ea88ee