Fix UB in NodeMapDelta
Summary:
If node map delta is passed 2 null maps we end up with some UB trying to call end() on a null object. MapDeltaImpl begin() here defers to end() which then ends up calling new_->end(). Crash signature P1410746766

The fix is to add special case when both new_ and old_ are null. In this case, create a sentinel Iterator with nothing in it. the result is begin() and end() both return this identical object so we won't do any iteration as we'd expect

Reviewed By: msomasundaran

Differential Revision:
D58485942

Privacy Context Container: L1125642

fbshipit-source-id: c36a598c94390b8c96987f9a569e9c0dd40fac5e